# swagger/doc.json
{
	"openapi": "3.0.1",
	"info": { "title": "API https://tvoe-dev.ru/", "version": "v1" },
	"servers": [{ "url": "http://localhost:4000" }],
...
	"paths": {
...
		"/profile/history": {
			"get": {
				"summary": "Профиль > История просмотров",
				"tags": ["Profile"],
				"parameters": [
					{
						"in": "query",
						"name": "limit",
						"schema": { "type": "number" },
						"description": "Ограничение на максимальное количество элементов в теле ответа"
					},
					{
						"in": "query",
						"name": "skip",
						"schema": { "type": "number" },
						"description": "Количествово пропускаемых элементов из начала выборки"
					}
				],
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"totalSize": {
											"type": "number",
											"description": "Общее количество сущностей в БД"
										},
										"items": {
											"type": "array",
											"uniqueItems": true,
											"minItems": 0,
											"items": {
												"type": "object",
												"properties": {
													"name": {
														"type": "string",
														"description": "Название фильма"
													},
													"shortDesc": {
														"type": "string",
														"description": "Краткое описание"
													},
													"trailer": {
														"type": "object",
														"properties": {
															"id": {
																"type": "string",
																"description": "Mongoose.ObjectId"
															},
															"src": {
																"type": "string",
																"description": "Путь к файлу изображению обложки"
															},
															"thumbnail": {
																"type": "string",
																"description": "Миниатюра (кадр из видео)"
															},
															"duration": {
																"type": "number",
																"description": "Длительность"
															}
														}
													},
													"ageLevel": {
														"type": "string",
														"description": "Возрастное ограничение"
													},
													"dateReleased": {
														"type": "Date",
														"description": "Дата релиза"
													},
													"categoryAlias": {
														"type": "string",
														"description": "Вид категориию. Может быть films, serials или collections"
													},
													"poster": {
														"type": "object",
														"properties": {
															"src": {
																"type": "string",
																"description": "Путь к файлу изображению"
															}
														}
													},
													"rating": {
														"type": "number",
														"description": "Рейтинг от 0 до 10"
													},
													"duration": {
														"type": "string",
														"description": "Длительность фильма в секундах"
													},
													"url": {
														"type": "string",
														"description": "Путь к фильму"
													},
													"moviepagelogs": {
														"type": "object",
														"properties": {
															"_id": {
																"type": "string",
																"description": "Id таблицы moviePageLogSchema (mongoose.Schema.Types.ObjectId)"
															},
															"count": {
																"type": "number",
																"description": "Kоличество просмотров фильма"
															},
															"updatedAt": {
																"type": "array",
																"items": {
																	"type": "Date",
																	"description": "Дата обновления"
																}
															}
														}
													},
													"lastLogUpdatedAt": {
														"type": "array",
														"items": {
															"type": "Date",
															"description": "Дата последнего обновления"
														}
													}
												}
											}
										}
									}
								},
								"examples": {
									"example": {
										"value": {
											"totalSize": 2,
											"items": [
												{
													"name": "Америка: фильм",
													"shortDesc": "Вооруженный пилами Джордж Вашингтон объединяется с обожающим пиво друганом Сэмом Адамсом, чтобы приструнить британцев в ироничной комедии о Войне за независимость США.",
													"trailer": {
														"_id": "641bf64b0073f02ab65dde80",
														"src": "/videos/641bf64b0073f02ab65dde7f.mp4",
														"thumbnail": "/images/641bf64b0073f02ab65dde81.jpg"
													},
													"ageLevel": 18,
													"dateReleased": "2021-06-15",
													"categoryAlias": "films",
													"poster": {
														"src": "/images/641f4080944e8bd192d18318.jpg"
													},
													"rating": 7,
													"duration": 5892,
													"url": "/p/amerika-film",
													"moviepagelogs": {
														"_id": "64007d7d0a569a75e14fc2a1",
														"count": 1,
														"updatedAt": [
															"2023-06-07T16:15:48.5030000Z"
														]
													},
													"lastLogUpdatedAt": ["2023-06-07T16:15:48.503Z"]
												},
												{
													"name": "Денежный выстрел: История Pornhub",
													"shortDesc": "Проникните глубже в историю самой известной платформы для взрослых Pornhub и недавний скандал вокруг неё. Детская порнография, сексуальное насилие, половая эксплуатация, торговля людьми — это лишь небольшой перечень того, что было найдено на страницах всем известного сайта",
													"trailer": {
														"_id": "6419a8a30073f02ab65d4f07",
														"src": "/videos/6419a8a30073f02ab65d4f06.mp4",
														"thumbnail": "/images/6419a8a30073f02ab65d4f08.jpg"
													},
													"ageLevel": 18,
													"dateReleased": "2023-02-17",
													"categoryAlias": "films",
													"poster": {
														"src": "/images/63fded0598e2af31cb68b46f.jpg"
													},
													"rating": 10,
													"duration": 7063.708333,
													"url": "/p/snyat-blokirovku",
													"moviepagelogs": {
														"_id": "6419a7540073f02ab65d4e36",
														"count": 1,
														"updatedAt": [
															"2023-06-07T16:15:48.5030000Z"
														]
													},
													"lastLogUpdatedAt": ["2023-06-07T16:15:48.503Z"]
												}
											]
										}
									}
								}
							}
						}
					}
				}
			}
		},
...
		"/admin/users/movies/movie": {
			"get": {
				"summary": "Админ-панель > Получение одной записи ( фильма или сериала )",
				"tags": ["Admin"],
				"parameters": [
					{
						"in": "query",
						"name": "skip",
						"schema": { "type": "string" },
						"description": "Кол-во пропускаемых элементов из начала выборки"
					},
					{
						"in": "query",
						"name": "limit",
						"schema": { "type": "string" },
						"description": "Ограничение на максимальное количество элементов в теле ответа"
					},
					{
						"in": "query",
						"name": "id",
						"schema": { "type": "string" },
						"description": "Название фильма или сериала"
					}
				],
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"_id": {
											"type": "string",
											"description": "Id фильма (mongoose.ObjectId)"
										},
										"name": {
											"type": "string",
											"description": "Название фильма на русском"
										},
										"origName": {
											"type": "string",
											"description": "Оригинальное название фильма на английском"
										},
										"shortDesc": {
											"type": "string",
											"description": "Краткое описание"
										},
										"fullDesc": {
											"type": "string",
											"description": "Полное описание"
										},
										"alias": {
											"type": "string",
											"description": "Сгенерированный человекопонятный url"
										},
										"ageLevel": {
											"type": "string",
											"description": "Возрастное ограничение"
										},
										"dateReleased": {
											"type": "string",
											"description": "Дата выпуска"
										},
										"countries": {
											"type": "array",
											"description": "Cписок стран",
											"uniqueItems": true,
											"minItems": 0,
											"items": {
												"type": "string",
												"description": "Название страны"
											}
										},
										"categoryAlias": {
											"type": "string",
											"description": "Вид категориию. Может быть films, serials"
										},
										"genresAliases": {
											"type": "array",
											"description": "Cписок жанров",
											"uniqueItems": true,
											"minItems": 0,
											"items": {
												"type": "string",
												"description": "Название жанра"
											}
										},
										"persons": {
											"type": "array",
											"description": "Список актёров и преставителей съемочной группы",
											"uniqueItems": true,
											"minItems": 0,
											"items": {
												"type": "object",
												"properties": {
													"_id": {
														"type": "string",
														"description": "mongoose.ObjectId"
													},
													"name": {
														"type": "string",
														"description": "Имя и фамилия человека"
													},
													"type": {
														"type": "string",
														"description": "director - режиссер, actor - актер, producer - продюсер, screenwriter - сценарист, operator - оператор"
													}
												}
											}
										},
										"series": {
											"type": "array",
											"description": "Список серий. Приходят в во вложенном массиве",
											"uniqueItems": true,
											"minItems": 0,
											"items": {
												"type": "array",
												"description": "Список серий",
												"uniqueItems": true,
												"minItems": 0,
												"items": {
													"type": "object",
													"properties": {
														"_id": {
															"type": "string",
															"description": "Id cерии (mongoose.ObjectId)"
														},
														"src": {
															"type": "string",
															"description": "источник на файл ( путь к исходнику )"
														},
														"duration": {
															"type": "number",
															"description": "Продолжительность серии"
														},
														"thumbnail": {
															"type": "string",
															"description": "миниатюра (кадр из видео)"
														}
													}
												}
											}
										},
										"films": {
											"type": "array",
											"description": "список фильмов",
											"uniqueItems": true,
											"minItems": 0,
											"items": {
												"type": "object",
												"properties": {
													"_id": {
														"type": "string",
														"description": "Имя и фамилия человека"
													},
													"src": {
														"type": "string",
														"description": "источник на файл ( путь к исходнику )"
													},
													"duration": {
														"type": "number",
														"description": "длительность фильма"
													},
													"thumbnail": {
														"type": "string",
														"description": "миниатюра (кадр из видео)"
													}
												}
											}
										},
										"createdAt": {
											"type": "Date",
											"description": "Дата создания"
										},
										"updatedAt": {
											"type": "Date",
											"description": "Дата обновления"
										},
										"poster": {
											"type": "object",
											"properties": {
												"id": {
													"type": "string",
													"description": "mongoose.ObjectId"
												},
												"src": {
													"type": "string",
													"description": "Путь к файлу изображению постера"
												}
											}
										},
										"cover": {
											"type": "object",
											"properties": {
												"id": {
													"type": "string",
													"description": "mongoose.ObjectId"
												},
												"src": {
													"type": "string",
													"description": "Путь к файлу изображению обложки"
												}
											}
										},
										"logo": {
											"type": "object",
											"properties": {
												"id": {
													"type": "string",
													"description": "mongoose.ObjectId"
												},
												"src": {
													"type": "string",
													"description": "Путь к файлу изображению логотипа"
												}
											}
										},
										"trailer": {
											"type": "object",
											"properties": {
												"id": {
													"type": "string",
													"description": "mongoose.ObjectId"
												},
												"src": {
													"type": "string",
													"description": "Путь к файлу изображению обложки"
												},
												"thumbnail": {
													"type": "string",
													"description": "миниатюра (кадр из видео)"
												},
												"duration": {
													"type": "number",
													"description": "Длительность",
													"required": false
												}
											}
										},
										"raisedUpAt": {
											"type": "Date",
											"description": "Дата поднятия в списке (для актуальности)"
										},
										"publishedAt": {
											"type": "Date",
											"description": "Дата публикации"
										}
									}
								},
								"examples": {
									"example": {
										"value": {
											"name": "Убийство в Париже",
											"countries": ["США"],
											"genresAliases": ["detektivy", "komedii", "kriminal"],
											"persons": [
												{
													"name": " Джереми Гарелик",
													"type": "director",
													"_id": "642f383a8e6fd86616fdf3a9"
												},
												{
													"name": "Адам Сэндлер",
													"type": "actor",
													"_id": "642f383a8e6fd86616fdf3aa"
												},
												{
													"name": "Дженнифер Энистон",
													"type": "actor",
													"_id": "642f383a8e6fd86616fdf3ab"
												},
												{
													"name": "Марк Стронг",
													"type": "actor",
													"_id": "642f383a8e6fd86616fdf3ac"
												},
												{
													"name": "Мелани Лоран",
													"type": "actor",
													"_id": "642f383a8e6fd86616fdf3ad"
												},
												{
													"name": "Джоди Тёрнер-Смит",
													"type": "actor",
													"_id": "642f383a8e6fd86616fdf3ae"
												},
												{
													"name": "Джон Кани",
													"type": "actor",
													"_id": "642f383a8e6fd86616fdf3af"
												},
												{
													"name": "Куху Верма",
													"type": "actor",
													"_id": "642f383a8e6fd86616fdf3b0"
												},
												{
													"name": "Дэни Бун",
													"type": "actor",
													"_id": "642f383a8e6fd86616fdf3b1"
												},
												{
													"name": "Адиль Ахтар",
													"type": "actor",
													"_id": "642f383a8e6fd86616fdf3b2"
												},
												{
													"name": "Энрике Арсе",
													"type": "actor",
													"_id": "642f383a8e6fd86616fdf3b3"
												},
												{
													"name": " Дженнифер Энистон",
													"type": "producer",
													"_id": "642f383a8e6fd86616fdf3b4"
												},
												{
													"name": "Рафаэль Бенолье",
													"type": "producer",
													"_id": "642f383a8e6fd86616fdf3b5"
												},
												{
													"name": "Бэрри Бернарди",
													"type": "producer",
													"_id": "642f383a8e6fd86616fdf3b6"
												},
												{
													"name": " Джеймс Вандербилт",
													"type": "screenwriter",
													"_id": "642f383a8e6fd86616fdf3b7"
												},
												{
													"name": " Боян Бацелли",
													"type": "operator",
													"_id": "642f383a8e6fd86616fdf3b8"
												}
											],
											"series": [],
											"films": [
												{
													"_id": "642f3a478e6fd86616fdf425",
													"src": "/videos/642f3a3c8e6fd86616fdf422.mp4",
													"width": 1920,
													"height": 1080,
													"duration": 5455.125333,
													"thumbnail": "/images/642f3a468e6fd86616fdf424.jpg"
												}
											],
											"createdAt": "2023-04-06T21:15:34.883Z",
											"updatedAt": "2023-04-20T17:34:04.416Z",
											"origName": "Murder Mystery",
											"shortDesc": "Ник и Одри Шпиц — теперь уже владельцы своего детективного агентства — расследуют важное для их карьеры дело о похищении друга-миллиардера с его собственной свадьбы.",
											"fullDesc": "Ник и Одри Шпиц — теперь уже владельцы своего детективного агентства — расследуют важное для их карьеры дело о похищении друга-миллиардера с его собственной свадьбы.",
											"dateReleased": "2023-03-31",
											"ageLevel": 12,
											"categoryAlias": "films",
											"logo": {
												"_id": "642f38428e6fd86616fdf3cf",
												"src": "/images/642f38428e6fd86616fdf3d0.png"
											},
											"trailer": {
												"_id": "642f38728e6fd86616fdf40d",
												"src": "/videos/642f38728e6fd86616fdf40a.mp4",
												"thumbnail": "/images/642f38728e6fd86616fdf40c.jpg"
											},
											"alias": "ubiystvo-v-parizhe",
											"publishedAt": "2023-04-20T17:34:04.414Z",
											"cover": {
												"_id": "642f587f8e6fd86616fe0773",
												"src": "/images/642f587f8e6fd86616fe0774.jpg"
											},
											"poster": {
												"_id": "642f5f6a8e6fd86616fe097d",
												"src": "/images/642f5f6a8e6fd86616fe097e.jpg"
											},
											"raisedUpAt": "2023-04-20T17:34:04.414Z"
										}
									}
								}
							}
						}
					}
				}
			}
		},
...
		"/admin/moviesRatingHistory": {
			"get": {
				"summary": "Админ-панель > История рейтингов",
				"tags": ["Admin"],
				"parameters": [
					{
						"in": "query",
						"name": "limit",
						"schema": { "type": "number" },
						"description": "Ограничение на максимальное количество элементов в теле ответа"
					},
					{
						"in": "query",
						"name": "skip",
						"schema": { "type": "number" },
						"description": "Кол-во пропускаемых элементов из начала выборки"
					},
					{
						"in": "query",
						"name": "query",
						"schema": { "type": "number" },
						"description": "Для поиска документов (сущностей коллекции users), с которыми совпадают символы в имени и емейлу. Либо для полного совпадения по id (mongoose.ObjectId)  пользователя"
					}
				],
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"totalSize": {
											"type": "number",
											"description": "Общее количество сущностей в БД (документов коллекции users)"
										},
										"items": {
											"type": "array",
											"uniqueItems": true,
											"minItems": 0,
											"items": {
												"type": "object",
												"properties": {
													"_id": {
														"type": "string",
														"description": "Id (mongoose.ObjectId) оценки фильма пользователем (документа коллекции MovieRating)"
													},
													"rating": {
														"type": "number",
														"description": "Оценка фильма пользователем"
													},
													"createdAt": {
														"type": "Date",
														"description": "Дата создания"
													},
													"updatedAt": {
														"type": "Date",
														"description": "Дата обновления"
													},
													"movie": {
														"type": "object",
														"properties": {
															"name": {
																"type": "string",
																"description": "Название фильма"
															},
															"poster": {
																"type": "object",
																"properties": {
																	"src": {
																		"type": "string",
																		"description": "Путь к постеру (файлу картинке)"
																	}
																}
															}
														}
													},
													"user": {
														"type": "object",
														"properties": {
															"_id": {
																"type": "string",
																"description": "Profile id"
															},
															"email": {
																"type": "string",
																"description": "Email user"
															},
															"avatar": {
																"type": "string",
																"description": "Path to img src profile user"
															},
															"firstname": {
																"type": "string",
																"description": "Имя этого человека"
															},
															"subscribe": {
																"type": "object",
																"description": "Информация по подписке. Если подписки нет, в этом поле null",
																"properties": {
																	"startAt": {
																		"type": "Date",
																		"description": "Дата начала подписки"
																	},
																	"finishAt": {
																		"type": "Date",
																		"description": "Дата окончания подписки"
																	},
																	"tariffId": {
																		"type": "string",
																		"description": "Id тарифа (mongoose.ObjectId)"
																	}
																}
															},
															"role": {
																"type": "string",
																"description": "Admin, manager или если это обычный пользователь то это поле отсуствует."
															}
														}
													}
												}
											}
										}
									},
									"example": {
										"value": {
											"totalSize": 189,
											"items": [
												{
													"_id": "6489f001a2871402bdd31351",
													"rating": 7,
													"createdAt": "2023-06-14T16:51:13.791Z",
													"updatedAt": "2023-06-14T16:51:13.791Z",
													"__v": 0,
													"movie": {
														"name": "Выстрел в сердце",
														"poster": {
															"src": "/images/641f44d1944e8bd192d18c57.jpg"
														}
													},
													"user": {
														"_id": "63bbed7523c814afb608a3f6",
														"email": "evlevelx@yandex.ru",
														"avatar": "/images/64808068904d20e5d4c1b555.jpg",
														"firstname": "Евгений",
														"subscribe": {
															"startAt": "2023-03-22T10:17:28.873Z",
															"finishAt": "2023-09-18T10:17:28.873Z",
															"tariffId": "63dbe1687f457bc81bc920cb"
														},
														"role": "admin"
													}
												},
												{
													"_id": "6489efa0a2871402bdd31003",
													"rating": 6,
													"createdAt": "2023-06-14T16:49:36.628Z",
													"updatedAt": "2023-06-14T16:49:36.628Z",
													"__v": 0,
													"movie": {
														"name": "Битва при Нарвике: Первое поражение Гитлера",
														"poster": {
															"src": "/images/640376060a569a75e1507a94.jpg"
														}
													},
													"user": {
														"_id": "63bbed7523c814afb608a3f6",
														"email": "evlevelx@yandex.ru",
														"avatar": "/images/64808068904d20e5d4c1b555.jpg",
														"firstname": "Евгений",
														"subscribe": {
															"startAt": "2023-03-22T10:17:28.873Z",
															"finishAt": "2023-09-18T10:17:28.873Z",
															"tariffId": "63dbe1687f457bc81bc920cb"
														},
														"role": "admin"
													}
												}
											]
										}
									}
								}
							}
						}
					}
				}
			}
		},
...
		"/movies": {
			"get": {
				"summary": "Список фильмов и сериалов",
				"tags": ["Movies"],
				"parameters": [
					{
						"in": "query",
						"name": "limit",
						"schema": { "type": "number" },
						"description": "Ограничение на максимальное количество элементов в теле ответа"
					},
					{
						"in": "query",
						"name": "skip",
						"schema": { "type": "number" },
						"description": "Количество пропускаемых элементов из начала выборки"
					}
				],
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"totalSize": {
											"type": "number",
											"description": "Общее количество сущностей(фильмов) в коллекции Movies"
										},
										"items": {
											"type": "array",
											"uniqueItems": true,
											"minItems": 0,
											"items": {
												"type": "object",
												"properties": {
													"_id": {
														"type": "string",
														"description": "Id фильма (mongoose.ObjectId)"
													},
													"aliasInUrl": {
														"type": "string",
														"description": "Вид категории. Может быть films, serials"
													},
													"alias": {
														"type": "string",
														"description": "Сгенерированный ЧПУ (человеко понятный url)"
													}
												}
											}
										}
									},
									"example": {
										"totalSize": 407,
										"items": [
											{
												"_id": "63ac7d6ab2c93ed221218186",
												"alias": "armiya-vorov",
												"category": { "aliasInUrl": "filmy" }
											},
											{
												"_id": "63ac8aa1b2c93ed2212186e7",
												"alias": "uendell-i-uayld",
												"category": { "aliasInUrl": "filmy" }
											},
											{
												"_id": "63acabd6b2c93ed221218e9e",
												"alias": "aferist-iz-tinder",
												"category": { "aliasInUrl": "filmy" }
											}
										]
									}
								}
							}
						}
					}
				}
			}
		},
...
		"/movies/rating": {
			"get": {
				"summary": "Получить рейтинг и комментарий поставленный пользователем",
				"tags": ["Movies"],
				"parameters": [
					{
						"in": "query",
						"name": "movieId",
						"schema": { "type": "string" },
						"description": "Id(mongoose.ObjectId) фильма (документа коллекции movie)",
						"required": true
					}
				],
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"rating": {
											"type": "number",
											"description": "Рейтинг который поставил пользователь фильму с переданным id"
										},
										"review": {
											"type": "string",
											"description": "Комментарий (отзыв) пользователя к фильму/cериалу"
										}
									},
									"example": { "rating": 8, "review": "Классный фильм" }
								}
							}
						}
					}
				}
			},
			"post": {
				"summary": "Отправка рейтинга(оценки) и отзыва(комментария)",
				"tags": ["Movies"],
				"requestBody": {
					"content": {
						"application/json": {
							"schema": {
								"type": "object",
								"properties": {
									"movieId": {
										"type": "string",
										"description": "Id фильма (mongoose.ObjectId)",
										"required": true
									},
									"rating": {
										"type": "number",
										"description": "Оценка фильма по шкале от 1 до 10"
									},
									"review": {
										"type": "string",
										"description": "Комментарий (отзыв) пользователя к фильму/cериалу"
									}
								}
							}
						}
					}
				},
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"success": {
											"type": "boolean",
											"description": "Значение true - сообщение об успехе"
										},
										"movieId": {
											"type": "string",
											"description": "Id фильма (mongoose.ObjectId)"
										},
										"rating": {
											"type": "number",
											"description": "Рейтинг который поставил пользователь фильму с переданным id"
										}
									},
									"example": {
										"success": true,
										"movieId": "63ac8aa1b2c93ed2212186e7",
										"rating": 9,
										"review": "Классный фильм"
									}
								}
							}
						}
					}
				}
			},
			"delete": {
				"summary": "Удаления рейтинга и отзыва",
				"tags": ["Movies"],
				"requestBody": {
					"content": {
						"application/json": {
							"schema": {
								"type": "object",
								"properties": {
									"movieId": {
										"type": "string",
										"description": "Id фильма (mongoose.ObjectId)",
										"required": true
									}
								}
							}
						}
					}
				},
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"type": {
											"type": "string",
											"description": "Надпись \"success\""
										},
										"alert": {
											"type": "boolean",
											"description": "Значение true"
										},
										"_id": {
											"type": "string",
											"description": "Id удаленного фильма или сериала (mongoose.ObjectId)"
										},
										"msg": {
											"type": "string",
											"description": "Надпись \"Успешно удалено\""
										}
									},
									"example": {
										"type": "success",
										"alert": true,
										"id": "63ac7d6ab2c93ed221218186",
										"msg": "Успешно удалено"
									}
								}
							}
						}
					}
				}
			}
		},
...
	},
...
}